+++ ChargingApp/DTOs/OrderAdminDto.cs
@@ -8,7 +8,7 @@ public class OrderAdminDto
     public string EnglishName { get; set; }
     public double TotalPrice { get; set; }
     public string? PlayerId { get; set; }
-    public int Quantity { get; set; } = 1;
+    public double Quantity { get; set; } = 1;
     public string OrderType { get; set; } = "Normal";
     public string Notes { get; set; }= "Pending";
 }
